+++ src/main/python/plotlyst/view/widget/character/profile.py
@@ -34,7 +34,8 @@ from qtmenu import MenuWidget, ActionTooltipDisplayMode
 
 from plotlyst.core.domain import Character, CharacterProfileSectionReference, CharacterProfileFieldReference, \
     CharacterProfileFieldType, CharacterMultiAttribute, CharacterProfileSectionType, MultiAttributePrimaryType, \
-    MultiAttributeSecondaryType, CharacterSecondaryAttribute, StrengthWeaknessAttribute
+    MultiAttributeSecondaryType, CharacterSecondaryAttribute, StrengthWeaknessAttribute, CharacterPersonalityAttribute
+from plotlyst.core.help import enneagram_help, mbti_keywords, mbti_help
 from plotlyst.core.template import TemplateField, iq_field, eq_field, rationalism_field, willpower_field, \
     creativity_field, traits_field, values_field, flaw_placeholder_field, goal_field, internal_goal_field, stakes_field, \
     conflict_field, motivation_field, methods_field, internal_motivation_field, internal_conflict_field, \
@@ -42,7 +43,8 @@ from plotlyst.core.template import TemplateField, iq_field, eq_field, rationalis
     baggage_coping_field, baggage_healing_field, baggage_deterioration_field, ghost_field, \
     baggage_defense_mechanism_field, wound_field, demon_field, baggage_trigger_field, fear_field, misbelief_field, \
     flaw_triggers_field, flaw_coping_field, flaw_manifestation_field, flaw_relation_field, flaw_goals_field, \
-    flaw_growth_field, flaw_deterioration_field
+    flaw_growth_field, flaw_deterioration_field, enneagram_choices, SelectionItem, mbti_choices, love_style_choices, \
+    work_style_choices
 from plotlyst.env import app_env
 from plotlyst.view.common import tool_btn, wrap, emoji_font, action, insert_before_the_end, push_btn, label, \
     fade_out_and_gc
@@ -50,10 +52,10 @@ from plotlyst.view.icons import IconRegistry
 from plotlyst.view.layout import group
 from plotlyst.view.style.slider import apply_slider_color
 from plotlyst.view.widget.button import CollapseButton, SecondaryActionPushButton, DotsMenuButton
-from plotlyst.view.widget.character.editor import StrengthWeaknessEditor
-from plotlyst.view.widget.display import Icon
+from plotlyst.view.widget.character.editor import StrengthWeaknessEditor, DiscSelector, EnneagramSelector, MbtiSelector, \
+    LoveStyleSelector
+from plotlyst.view.widget.display import Icon, Emoji, dash_icon
 from plotlyst.view.widget.input import AutoAdjustableTextEdit, Toggle, TextInputDialog
-from plotlyst.view.widget.progress import CircularProgressBar
 from plotlyst.view.widget.template.impl import TraitSelectionWidget, LabelsSelectionWidget
 
 
@@ -1190,6 +1192,242 @@ class StrengthsWeaknessesFieldWidget(TemplateFieldWidgetBase):
         self.valueFilled.emit(value / count if count else 0)
 
 
+class EnneagramFieldWidget(TemplateFieldWidgetBase):
+    def __init__(self, character: Character, parent=None):
+        super(EnneagramFieldWidget, self).__init__(parent)
+        self.character = character
+        self.wdgEditor = EnneagramSelector()
+        self._defaultTooltip: str = 'Select Enneagram personality'
+        _layout = vbox(self)
+        _layout.addWidget(self.wdgEditor, alignment=Qt.AlignmentFlag.AlignTop)
+
+        emojiDesire = Emoji()
+        emojiDesire.setText(emoji.emojize(':smiling_face:'))
+        emojiDesire.setToolTip('Core desire')
+        emojiFear = Emoji()
+        emojiFear.setText(emoji.emojize(':face_screaming_in_fear:'))
+        emojiFear.setToolTip('Core fear')
+        self.lblDesire = QLabel('')
+        self.lblDesire.setToolTip('Core desire')
+        self.lblFear = QLabel('')
+        self.lblFear.setToolTip('Core fear')
+
+        decr_font(emojiDesire, 2)
+        decr_font(self.lblDesire)
+        decr_font(emojiFear, 2)
+        decr_font(self.lblFear)
+
+        self.wdgAttr = group(
+            group(dash_icon(), emojiDesire, self.lblDesire, spacer()),
+            group(dash_icon(), emojiFear, self.lblFear, spacer()),
+            vertical=False)
+        margins(self.wdgAttr, left=10)
+        _layout.addWidget(self.wdgAttr)
+        self.wdgAttr.setHidden(True)
+
+        _layout.addWidget(spacer())
+
+        self.wdgEditor.selected.connect(self._selectionChanged)
+        self.wdgEditor.ignored.connect(self._ignored)
+
+        if self.character.personality.enneagram:
+            self.setValue(self.character.personality.enneagram.value)
+
+    @overrides
+    def value(self) -> Any:
+        return self.wdgEditor.value()
+
+    @overrides
+    def setValue(self, value: Any):
+        self.wdgEditor.setValue(value)
+        enneagram = enneagram_choices.get(value)
+        if enneagram:
+            # self.wdgEditor.setToolTip(enneagram_help[value])
+            self._selectionChanged(enneagram)
+        elif value is None:
+            self._ignored()
+        else:
+            self.wdgEditor.setToolTip(self._defaultTooltip)
+
+    def _selectionChanged(self, item: SelectionItem):
+        if self.character.personality.enneagram is None:
+            self.character.personality.enneagram = CharacterPersonalityAttribute()
+        self.character.personality.enneagram.value = item.text
+
+        self.lblDesire.setText(item.meta['desire'])
+        self.lblFear.setText(item.meta['fear'])
+        self.wdgEditor.setToolTip(enneagram_help[item.text])
+        if self.isVisible():
+            qtanim.fade_in(self.wdgAttr)
+        else:
+            self.wdgAttr.setVisible(True)
+
+        self.valueFilled.emit(1)
+
+    def _ignored(self):
+        self.wdgEditor.setToolTip('Enneagram field is ignored for this character')
+        self.lblDesire.setText('')
+        self.lblFear.setText('')
+        self.wdgAttr.setHidden(True)
+        self.valueFilled.emit(1)
+
+
+class MbtiFieldWidget(TemplateFieldWidgetBase):
+    def __init__(self, character: Character, parent=None):
+        super(MbtiFieldWidget, self).__init__(parent)
+        self.character = character
+        self.wdgEditor = MbtiSelector()
+        self._defaultTooltip: str = 'Select MBTI personality type'
+        self.wdgEditor.setToolTip(self._defaultTooltip)
+
+        _layout = vbox(self)
+        _layout.addWidget(self.wdgEditor)
+
+        self.lblKeywords = label(wordWrap=True)
+        decr_font(self.lblKeywords)
+
+        self.wdgAttr = group(dash_icon(), self.lblKeywords, spacer())
+        margins(self.wdgAttr, left=10)
+        _layout.addWidget(self.wdgAttr)
+        self.wdgAttr.setHidden(True)
+
+        _layout.addWidget(spacer())
+
+        self.wdgEditor.selected.connect(self._selectionChanged)
+        self.wdgEditor.ignored.connect(self._ignored)
+
+        if self.character.personality.mbti:
+            self.setValue(self.character.personality.mbti.value)
+
+    @overrides
+    def value(self) -> Any:
+        return self.wdgEditor.value()
+
+    @overrides
+    def setValue(self, value: Any):
+        self.wdgEditor.setValue(value)
+        if value:
+            mbti = mbti_choices[value]
+            self._selectionChanged(mbti)
+        elif value is None:
+            self._ignored()
+        else:
+            self.wdgEditor.setToolTip(self._defaultTooltip)
+
+    def _selectionChanged(self, item: SelectionItem):
+        if self.character.personality.mbti is None:
+            self.character.personality.mbti = CharacterPersonalityAttribute()
+        self.character.personality.mbti.value = item.text
+
+        self.lblKeywords.setText(mbti_keywords.get(item.text, ''))
+        if self.isVisible():
+            qtanim.fade_in(self.wdgAttr)
+        else:
+            self.wdgAttr.setVisible(True)
+
+        self.wdgEditor.setToolTip(mbti_help[item.text])
+        self.valueFilled.emit(1)
+
+    def _ignored(self):
+        self.wdgEditor.setToolTip('MBTI field is ignored for this character')
+        self.wdgAttr.setHidden(True)
+        self.valueFilled.emit(1)
+
+
+class LoveStyleFieldWidget(TemplateFieldWidgetBase):
+    def __init__(self, character: Character, parent=None):
+        super().__init__(parent)
+        self.character = character
+        self.wdgEditor = LoveStyleSelector()
+        self._defaultTooltip: str = 'Select love style'
+        _layout = vbox(self)
+        _layout.addWidget(self.wdgEditor, alignment=Qt.AlignmentFlag.AlignLeft)
+
+        self.wdgEditor.selected.connect(self._selectionChanged)
+        self.wdgEditor.ignored.connect(self._ignored)
+
+        if self.character.personality.love:
+            self.setValue(self.character.personality.love.value)
+
+    @overrides
+    def value(self) -> Any:
+        return self.wdgEditor.value()
+
+    @overrides
+    def setValue(self, value: Any):
+        self.wdgEditor.setValue(value)
+        if value:
+            mbti = love_style_choices[value]
+            self._selectionChanged(mbti)
+        elif value is None:
+            self._ignored()
+        else:
+            self.wdgEditor.setToolTip(self._defaultTooltip)
+
+    def _selectionChanged(self, item: SelectionItem):
+        if self.character.personality.love is None:
+            self.character.personality.love = CharacterPersonalityAttribute()
+        self.character.personality.love.value = item.text
+
+    def _ignored(self):
+        self.wdgEditor.setToolTip('Love style field is ignored for this character')
+        self.valueFilled.emit(1)
+
+
+class WorkStyleFieldWidget(TemplateFieldWidgetBase):
+    def __init__(self, character: Character, parent=None):
+        super().__init__(parent)
+        self.character = character
+        self.wdgEditor = DiscSelector()
+        self._defaultTooltip: str = 'Select work style'
+        _layout = vbox(self)
+        _layout.addWidget(self.wdgEditor, alignment=Qt.AlignmentFlag.AlignLeft)
+
+        self.wdgEditor.selected.connect(self._selectionChanged)
+        self.wdgEditor.ignored.connect(self._ignored)
+
+        if self.character.personality.work:
+            self.setValue(self.character.personality.work.value)
+
+    @overrides
+    def value(self) -> Any:
+        return self.wdgEditor.value()
+
+    @overrides
+    def setValue(self, value: Any):
+        self.wdgEditor.setValue(value)
+        if value:
+            mbti = work_style_choices[value]
+            self._selectionChanged(mbti)
+        elif value is None:
+            self._ignored()
+        else:
+            self.wdgEditor.setToolTip(self._defaultTooltip)
+
+    def _selectionChanged(self, item: SelectionItem):
+        if self.character.personality.work is None:
+            self.character.personality.work = CharacterPersonalityAttribute()
+        self.character.personality.work.value = item.text
+
+    def _ignored(self):
+        self.wdgEditor.setToolTip('Work style field is ignored for this character')
+        self.valueFilled.emit(1)
+
+
+class PersonalityFieldWidget(TemplateFieldWidgetBase):
+    def __init__(self, character: Character, parent=None):
+        super().__init__(parent)
+        self.character = character
+
+        self._layout: QGridLayout = grid(self)
+        # margins(self, left=15)
+
+        self._layout.addWidget(EnneagramFieldWidget(self.character), 0, 0)
+        self._layout.addWidget(MbtiFieldWidget(self.character), 0, 1)
+        self._layout.addWidget(LoveStyleFieldWidget(self.character), 1, 0)
+        self._layout.addWidget(WorkStyleFieldWidget(self.character), 1, 1)
+
+
 def field_widget(ref: CharacterProfileFieldReference, character: Character) -> ProfileFieldWidget:
     if ref.type == CharacterProfileFieldType.Field_Summary:
         return SummaryField(character)
@@ -1207,6 +1445,8 @@ def field_widget(ref: CharacterProfileFieldReference, character: Character) -> P
         else:
             raise ValueError(f'Unrecognized field type {ref.type}')
         return FacultyField(ref, field, character)
+    elif ref.type == CharacterProfileFieldType.Field_Personality:
+        return PersonalityFieldWidget(character)
     elif ref.type == CharacterProfileFieldType.Field_Traits:
         return TraitsFieldWidget(character)
     elif ref.type == CharacterProfileFieldType.Field_Values:
